WIP: Replaced MAXIMUM_NUMBER_OF_PIECES with CAN_HANDLE_PIECE_REQUEST.
MAXIMUM_NUMBER_OF_PIECES had reduced to being a boolean. 1 for
serial sources, -1 for parallel sources. I removed it and added
a CAN_HANDLE_PIECE_REQUEST instead. This key, produced by a source,
tells the executive that a source is able to handle piece request.
It is a source only key produced in RequestInformation and is
not propagated downstream. If this key is not present, the executive
will only execute the source for piece 0 to produce the entire
data. It is then up to the user to add a filter that splits
the data for other piece requests. The only exception to this is
when CAN_PRODUCE_SUB_EXTENT is present, in which case the executive
will split using an extent translator AT THE source - not downstream.